# web/openapi.yaml
openapi: 3.1.0
info:
  title: Owlette API
  version: 2.3.1
  description: |
    The Owlette API lets you programmatically manage your fleet of Windows machines,
    deploy software, monitor processes, execute tools, and integrate with external systems
    via webhooks.

    ## Authentication

    All API requests require authentication via an **API key**. You can create API keys
    from the Owlette dashboard under **Settings → API Keys**, or via the
    [Create API Key](#tag/API-Keys/operation/createApiKey) endpoint.

    Include your API key in every request using one of these methods:

    - **Header** (recommended): `x-api-key: owk_your_key_here`
    - **Query parameter**: `?api_key=owk_your_key_here`

    API keys are scoped to your user account and inherit your site access permissions.

    ## Rate Limiting

    All endpoints are rate-limited. When you exceed the limit, you'll receive a `429 Too Many Requests`
    response with a `Retry-After` header indicating how long to wait.

    Default limit: **300 requests/hour** per API key.

    Response headers on every request:
    - `X-RateLimit-Limit` — Maximum requests per window
    - `X-RateLimit-Remaining` — Requests remaining in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    ## Errors

    All errors return a JSON object with an `error` field:

    ```json
    { "error": "Description of what went wrong" }
    ```

    Common HTTP status codes:
    | Code | Meaning |
...

...

...
  license:
...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /api/roosts:
    x-stub: true
    get:
      operationId: listRoosts
      tags: [roosts]
      summary: list roosts in a site
      description: |
        Lists roosts the caller has access to within a site.
        Cursor-paginated; `siteId` query is required.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      parameters:
        - $ref: '#/components/parameters/RoostVersionHeader'
        - $ref: '#/components/parameters/SiteIdQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/PageTokenQuery'
      responses:
        '200':
          description: Paginated roost list.
          headers:
            RateLimit-Limit:    { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset:    { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                type: object
                required: [items, next_page_token]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RoostSummary' }
                  next_page_token:
                    type: string
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }
    post:
      operationId: createRoost
      tags: [roosts]
      summary: create roost
      description: |
        Creates a new empty roost. Publish a version to populate it.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      parameters:
        - $ref: '#/components/parameters/RoostVersionHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoostRequest' }
      responses:
        '201':
          description: Roost created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Roost' }
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '402': { $ref: '#/components/responses/Problem402' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '409': { $ref: '#/components/responses/Problem409' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }
  /api/roosts/{roostId}:
    parameters:
      - $ref: '#/components/parameters/RoostIdPath'
      - $ref: '#/components/parameters/RoostVersionHeader'
    get:
      operationId: getRoost
      tags: [roosts]
      summary: fetch roost detail
      description: |
        Returns the roost's current state including `currentVersionId`
        and `previousVersionId`.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      x-stub: true
      parameters:
        - $ref: '#/components/parameters/SiteIdQuery'
      responses:
        '200':
          description: Roost detail.
          headers:
            RateLimit-Limit:    { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset:    { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Roost' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }
    patch:
      operationId: patchRoost
      tags: [roosts]
      summary: rename roost or change target machines
      description: |
        Updates the roost's name and/or target machine list. Does not
        affect version history.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      x-stub: true
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - $ref: '#/components/parameters/IfMatchHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchRoostRequest' }
      responses:
        '200':
          description: Roost updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Roost' }
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '409': { $ref: '#/components/responses/Problem409' }
        '412': { $ref: '#/components/responses/Problem412' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }
    delete:
      operationId: deleteRoost
      tags: [roosts]
      summary: soft-delete roost
      description: |
        Deletes the roost doc + versions subcollection. Content chunks
        in R2 are reclaimed by the chunk gc mark-and-sweep (not
        immediate). 30-day tombstone; chunks referenced only by this
        roost become eligible for gc at that point.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      parameters:
        - $ref: '#/components/parameters/SiteIdQuery'
      responses:
        '200':
          description: Roost deleted.
          content:
            application/json:
              schema:
                type: object
                required: [deleted, roostId, versionsDeleted]
                properties:
                  deleted:
                    type: boolean
                    example: true
                  roostId:
                    type: string
                  versionsDeleted:
                    type: integer
                    description: Count of version docs purged.
        '400': { $ref: '#/components/responses/Problem400' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '412': { $ref: '#/components/responses/Problem412' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }

  # ──────────────────────────────────────────────────────────────────────
  # 3. versions — list + publish (POST) are live above. detail, files,
  # diff endpoints are proposed.
  # ──────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: x-api-key
      in: header
      description: |
        Legacy api key header — supported for backwards compatibility.
        New integrations should use `bearerApiKey` instead
        (`Authorization: Bearer owk_live_...`).
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: owk_live_* / owk_test_*
      description: |
        Preferred auth mechanism for the public roost api. Send the
        scoped api key as `Authorization: Bearer owk_live_<secret>`
        (production) or `Authorization: Bearer owk_test_<secret>`
        (sandbox). See [authentication.md](../docs/api/authentication.md).
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: |
        Firebase ID token issued to the authenticated user or agent device.
        Used by dashboard + agent requests to roost endpoints. First-party
        clients only — external integrations use `apiKey` /
        `bearerApiKey` instead.
  parameters:
    SiteIdQuery:
      name: siteId
      in: query
      required: true
      schema:
        type: string
      description: The site ID
...
    RoostIdPath:
      name: roostId
      in: path
      required: true
      schema:
        type: string
      description: The roost ID (deploy unit).
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: |
        Client-generated idempotency key. If the server has seen the same
        key on a request with the same body within the replay window, it
        returns the original response instead of re-executing.
    RoostVersionHeader:
      name: Roost-Version
      in: header
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: '2026-04-22'
      description: |
        Selects a dated api version. Optional in wave 1; required once
        a second version ships. See [`GET /api/version`](#tag/utility/operation/getApiVersion).
    IfMatchHeader:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
      description: |
        Optimistic concurrency token. Pass the prior resource version
        (e.g. a `currentVersionId` or opaque etag) to have the server
        reject the mutation with 412 if the resource has changed.
    PageSizeQuery:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: Maximum items per page.
    PageTokenQuery:
      name: page_token
      in: query
      required: false
      schema:
        type: string
      description: |
        Opaque cursor from the prior response's `next_page_token`. An
        empty string means no more pages.
...
  headers:
    RateLimitLimit:
      description: Window limit for the rate-limit bucket that fired for this request.
      schema:
        type: integer
    RateLimitRemaining:
      description: Calls remaining in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: Seconds until the current window fully resets.
      schema:
        type: integer
    RetryAfter:
      description: Seconds the client should wait before retrying (429 only).
      schema:
        type: integer
    RoostRateLimitedReason:
      description: |
        Which rate-limit bucket tripped (`global-rate`, `endpoint-rate`,
        `key-rate`, `site-concurrency`). Present only on 429.
      schema:
        type: string
        enum: [global-rate, endpoint-rate, key-rate, site-concurrency]
  responses:
    Problem400:
      description: Validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem401:
      description: Authentication required.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem402:
      description: Quota exceeded.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem403:
      description: Forbidden (e.g. cross-site access).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem404:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem409:
      description: Conflict (e.g. no previous version to roll back to).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem412:
      description: Precondition failed (optimistic concurrency miss on finalize).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem429:
      description: Rate limited.
      headers:
        RateLimit-Limit:           { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining:       { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset:           { $ref: '#/components/headers/RateLimitReset' }
        Retry-After:               { $ref: '#/components/headers/RetryAfter' }
        Roost-Rate-Limited-Reason: { $ref: '#/components/headers/RoostRateLimitedReason' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem500:
      description: Internal server error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
...
    ProblemDetails:
      type: object
      description: |
        RFC 7807 problem+json envelope returned by all roost endpoints on
        error. Unlike the legacy `ErrorResponse`, this carries a stable
        machine-readable `type` URI, a correlation `requestId`, and
        optional field-level `errors`.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the problem category.
          example: https://owlette.app/problems/validation-failed
        title:
          type: string
          description: Short, stable human-readable summary.
          example: validation failed
        status:
          type: integer
          description: HTTP status code, mirrors the response status.
          example: 400
        detail:
          type: string
          description: Occurrence-specific explanation; safe to show to users.
        instance:
          type: string
          description: URI reference for this occurrence (often the route path).
        requestId:
          type: string
          description: Correlation ID for trace lookup; quote in support tickets.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: |
            Field-level errors keyed by dotted JSON path (400/422 responses).
            Example: `{ "body.hashes": ["must be a non-empty array"] }`
...
    RoostSummary:
      type: object
      required: [id, siteId, name, targets, totalSize, totalFiles, updatedAt]
      properties:
        id:       { type: string }
        siteId:   { type: string }
        name:     { type: string }
        targets:
          type: array
          items: { type: string }
        currentVersionId:
          type: string
          nullable: true
          pattern: '^sha256:[0-9a-f]{64}$'
        previousVersionId:
          type: string
          nullable: true
          pattern: '^sha256:[0-9a-f]{64}$'
        totalSize:
          type: integer
          format: int64
          minimum: 0
        totalFiles:
          type: integer
          minimum: 0
        updatedAt:
          type: string
          format: date-time
    Roost:
      description: Full roost detail. Extends `RoostSummary` with `createdAt`.
      allOf:
        - $ref: '#/components/schemas/RoostSummary'
        - type: object
          required: [createdAt]
          properties:
            createdAt:
              type: string
              format: date-time
    CreateRoostRequest:
      type: object
      required: [siteId, name]
      properties:
        siteId:  { type: string }
        name:
          type: string
          maxLength: 128
          description: Display name. Must be unique within the site.
        targets:
          type: array
          items: { type: string }
          description: Machine ids to target; may be empty at create time.
    PatchRoostRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
        targets:
          type: array
          items: { type: string }

    # versions
...